requirements.txt
gitpython==1.0.1